+++ src/CubicAIExplorer/Models/FileListGroupMode.cs
@@ -1,0 +1,10 @@
+namespace CubicAIExplorer.Models;
+
+public enum FileListGroupMode
+{
+    None,
+    Name,
+    Type,
+    Size,
+    DateModified
+}