+++ src/CubicAIExplorer/Models/FileListSortMode.cs
@@ -1,0 +1,14 @@
+namespace CubicAIExplorer.Models;
+
+public enum FileListSortMode
+{
+    Name,
+    Size,
+    Type,
+    DateModified,
+    Company,
+    Version,
+    Dimensions,
+    Duration,
+    Manual
+}